requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.2
pydantic>=2.8.2
pydantic-settings>=2.4.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
authlib>=1.3.1
httpx>=0.27.0

# AI / LLM + Embeddings
openai>=1.51.0

# Agent orchestration
langchain>=0.3.0
langgraph>=0.2.0
langchain-text-splitters>=0.3.0

# Retrieval
rank-bm25>=0.2.2
chromadb>=0.5.5
numpy>=1.26.4

# Document parsing
pdfplumber>=0.11.0
python-docx>=1.1.0

# Storage
minio>=7.2.8

# Email
sendgrid>=6.11.0

# Cache / Queue
redis[hiredis]>=5.0.8
celery[redis]>=5.4.0

# Utils
structlog>=24.4.0
python-dotenv>=1.0.1
